Coarse-pace the protocol loop in parked wait-for-operator states
A completed feed hold, a parked safety door (ajar or closed, awaiting
cycle-start), and sleep carry no active motion, so they now take the
coarse idle poll instead of the 200 us segment-production pace. The
motion sub-phases stay tight-paced: a decelerating hold (Hold_Pending)
and a door retract or resume (Parking_Retracting/Resuming) still run
segments, verified by a hold-then-resume preserving position. Fixes a
protocol thread that pinned ~28% of the single-core i.MX6 whenever the
machine was left parked in Door or Hold.

Also fix laser_arm_test.c's gfio_rd_attr stub to take size_t, not
unsigned long, so it cross-compiles on the 32-bit ARM target where the
two types differ (x86 CI did not catch it).

+++ tests/laser_arm_test.c
@@ -24,6 +24,7 @@
   post-wait re-check.
 */
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdio.h>
 #include <string.h>
 
@@ -51,7 +52,7 @@ void gfcool_laser_armed(bool armed) { (void)armed; }
 void gf_stream_laser(unsigned char power, bool fire) { (void)power; (void)fire; }
 void gf_stream_laser_arm(bool armed) { stream_armed = armed; }
 void gf_stream_laser_latch(bool lock) { latch_locked_last = lock; }
-int  gfio_rd_attr(const char *a, char *b, unsigned long l)
+int  gfio_rd_attr(const char *a, char *b, size_t l)
 { (void)a; if (l) b[0] = '\0'; return -1; }
 float gfio_conf_read_float(const char *k, float fb) { (void)k; return fb; }
 void serial_poll(void) {}
